requirements.txt
discord.py
python-dotenv
Flask==3.0.2